feat: linked info of showtime for current date
